move headers that are to become semi-public
