+++ Makefile.Watcom
@@ -43,7 +43,8 @@ CP = copy
 
 CFLAGS = -3r -mf -hc -zff -zgf -zq -zm -zc -s -fr=con -w2 -fpi -oilrtfm -aa   &
          -wcd=201 -bt=nt -d+ -dCARES_BUILDING_LIBRARY &
-         -dNTDDI_VERSION=0x06020000 -I. -I.\include -I.\src\lib $(SYS_INCL)
+         -dNTDDI_VERSION=0x06020000 -I. -I.\include -I.\src\lib -I.\src\lib\include &
+         $(SYS_INCL)
 
 LFLAGS = option quiet, map, caseexact, eliminate
