+++ src/lib/CMakeLists.txt
@@ -53,6 +53,7 @@ IF (CARES_SHARED)
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>"
 		       "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>"
 		PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}"
+		        "${CMAKE_CURRENT_SOURCE_DIR}/include"
 	)
 
 	TARGET_COMPILE_DEFINITIONS (${PROJECT_NAME} PRIVATE HAVE_CONFIG_H=1 CARES_BUILDING_LIBRARY)
@@ -110,6 +111,7 @@ IF (CARES_STATIC)
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>"
 		       "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>"
 		PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}"
+		        "${CMAKE_CURRENT_SOURCE_DIR}/include"
 	)
 
 	TARGET_COMPILE_DEFINITIONS (${LIBNAME} PRIVATE HAVE_CONFIG_H=1 CARES_BUILDING_LIBRARY)
