+++ src/lib/Makefile.am
@@ -11,7 +11,8 @@ ACLOCAL_AMFLAGS = -I m4 --install
 AM_CPPFLAGS += -I$(top_builddir)/include \
                -I$(top_builddir)/src/lib \
                -I$(top_srcdir)/include \
-               -I$(top_srcdir)/src/lib
+               -I$(top_srcdir)/src/lib \
+               -I$(top_srcdir)/src/lib/include
 
 lib_LTLIBRARIES = libcares.la
