+++ src/tools/CMakeLists.txt
@@ -11,6 +11,7 @@ IF (CARES_BUILD_TOOLS)
 		PUBLIC "$<BUILD_INTERFACE:${PROJECT_BINARY_DIR}>"
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>"
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/lib>"
+		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/lib/include>"
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>"
 		       "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>"
 		PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}"
@@ -39,6 +40,7 @@ IF (CARES_BUILD_TOOLS)
 		PUBLIC "$<BUILD_INTERFACE:${PROJECT_BINARY_DIR}>"
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>"
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/lib>"
+		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/lib/include>"
 		       "$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>"
 		       "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>"
 		PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}"
