+++ src/tools/Makefile.am
@@ -16,6 +16,7 @@ AM_CPPFLAGS += -I$(top_builddir)/include \
                -I$(top_builddir)/src/lib \
                -I$(top_srcdir)/include \
                -I$(top_srcdir)/src/lib \
+               -I$(top_srcdir)/src/lib/include \
                -DCARES_NO_DEPRECATED
 
 include Makefile.inc
