+++ test/Makefile.am
@@ -2,12 +2,13 @@
 # SPDX-License-Identifier: MIT
 AUTOMAKE_OPTIONS = foreign subdir-objects nostdinc 1.9.6
 
-AM_CPPFLAGS += -I$(top_builddir)/include \
-               -I$(top_builddir)/src/lib \
-               -I$(top_srcdir)/include   \
-               -I$(top_srcdir)/src/lib   \
-               -I$(top_srcdir)/test      \
-               -I$(top_builddir)/test    \
+AM_CPPFLAGS += -I$(top_builddir)/include       \
+               -I$(top_builddir)/src/lib       \
+               -I$(top_srcdir)/include         \
+               -I$(top_srcdir)/src/lib         \
+               -I$(top_srcdir)/src/lib/include \
+               -I$(top_srcdir)/test            \
+               -I$(top_builddir)/test          \
                -I$(top_builddir) -DCARES_NO_DEPRECATED
 
 AM_CXXFLAGS :=
