+++ test/Makefile.m32
@@ -19,7 +19,7 @@ ARES_SRC_DIR = ..
 # Where to find the built c-ares static library
 ARES_BLD_DIR = ..
 ARESLIB = $(ARES_BLD_DIR)/src/lib/libcares.a
-CPPFLAGS = -I$(ARES_SRC_DIR)/include -I$(ARES_SRC_DIR)/src/lib -I$(GTEST_ROOT)/include -DCARES_STATICLIB -DCARES_NO_DEPRECATED
+CPPFLAGS = -I$(ARES_SRC_DIR)/include -I$(ARES_SRC_DIR)/src/lib -I$(ARES_SRC_DIR)/src/lib/include -I$(GTEST_ROOT)/include -DCARES_STATICLIB -DCARES_NO_DEPRECATED
 CXXFLAGS = -Wall $(PTHREAD_CFLAGS) -std=gnu++14
 LDFLAGS =
 LDLIBS = -lws2_32 -liphlpapi
